requirements.txt
Pillow===9.3.0
boto3
runpod
rembg